both integration tests in one
