+++ test/cordova-integration-test.html
@@ -25,7 +25,7 @@
 		(function(b,r,a,n,c,h,_,s,d,k){if(!b[n]||!b[n]._q){for(;s<_.length;)c(h,_[s++]);d=r.createElement(a);d.async=1;d.src="../dist/cordova/build.js";k=r.getElementsByTagName(a)[0];k.parentNode.insertBefore(d,k);b[n]=h}})(window,document,"script","branch",function(b,r){b[r]=function(){b._q.push([r,arguments])}},{_q:[],_v:1},"init data addListener removeListener setIdentity logout track link sendSMS referrals credits creditHistory applyCode validateCode getCode redeem banner closeBanner".split(" "),0);
 	</script>
 
-	<script type="text/javascript" src="4_cordova_integration.js"></script>
+	<script type="text/javascript" src="4_integration.js"></script>
 
 	<script type="text/javascript" src="saucelabs.js"></script>
