+++ test/cordova-integration-test.template.html
@@ -25,7 +25,7 @@
 		// INSERT INIT CODE
 	</script>
 
-	<script type="text/javascript" src="4_cordova_integration.js"></script>
+	<script type="text/javascript" src="4_integration.js"></script>
 
 	<script type="text/javascript" src="saucelabs.js"></script>
