+++ test/integration-test.html
@@ -9,7 +9,7 @@
 	<link rel="stylesheet" type="text/css" href="../node_modules/mocha/mocha.css" />
 	<script src="../compiler/library/closure-library-master/closure/goog/base.js"></script>
 	<script>
-		goog.addDependency("../../../../../src/0_config.js", ['config'], []);
+		goog.addDependency("../../../../../test/web-config.js", ['config'], []);
 	</script>
 	<script src="../node_modules/mocha/mocha.js"></script>
 	<script src="../node_modules/sinon/pkg/sinon-1.13.0.js"></script>
