+++ test/web-config.js
@@ -5,7 +5,7 @@ goog.provide('config');
 
 config.link_service_endpoint = 'https://bnc.lt';
 config.api_endpoint = 'https://api.branch.io';
-config.version = '0.0.0';
+config.version = '1.5.8';
 
 /** @define {boolean} */
 window.WEB_BUILD = true;
